% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template.R
\name{read_template}
\alias{read_template}
\title{Read a template from a csv file}
\usage{
read_template(file)
}
\arguments{
\item{file}{A character vector with the path to the csv file}
}
\value{
a data.frame with the content of the csv file
}
\description{
The template should be created using \code{create_template()}
}
\examples{
# Create a temporary file for this example
file <- tempfile()

# Create a template and save it to a csv file
create_template(authors = c("Josep Maria", "Jane Doe"), file = file)

# Read the template back (in real life once it has been populated)
read_template(file)
}
