% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_read.R
\name{template_read}
\alias{template_read}
\title{Read a template from a csv file}
\usage{
template_read(file)
}
\arguments{
\item{file}{A character vector with the path to the csv file}
}
\value{
a data.frame with the content of the csv file
}
\description{
The template should be created using \code{create_template()}
}
\examples{
# Create a temporary file for this example
file <- tempfile()

# Create a template and save it to a csv file
template_create(authors = c("Josep Maria", "Jane Doe"), file = file)

# Read the template back (in real life once it has been populated)
template_read(file)
}
