% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cras.R
\name{write_cras}
\alias{write_cras}
\title{Write CRediT author statement}
\usage{
write_cras(cras_table, file = "cras.txt", overwrite = FALSE, markdown = FALSE)
}
\arguments{
\item{cras_table}{A data.frame created using \code{create_template()}}

\item{file}{The text file to be created}

\item{overwrite}{If TRUE, the file is overwritten. Otherwise, a error is triggered.}

\item{markdown}{If TRUE, the authors are surrounded by ** to make them bold in markdown.}
}
\value{
A text file with the CRedit authors statement
}
\description{
Write CRediT author statement
}
\examples{
# Generate a template and populate it (randomwly for this example)
cras_table <- create_template(authors = c("Josep Maria", "Jane Doe"))
cras_table[,2:ncol(cras_table)] <- sample(0:1, (ncol(cras_table)-1)*2, replace = TRUE)

# Create a temporary file just for this example
file <- tempfile()

# Write to the file
write_cras(cras_table, file, markdown = TRUE)

# Check the content of the file
readLines(file)
}
